[CLEANUP] Move utility classes to Pelago\Emogrifier\Utilities namespace
Fixes #711
