+++ src/Emogrifier/Utilities/CssConcatenator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pelago\Emogrifier;
+namespace Pelago\Emogrifier\Utilities;
 
 /**
  * Facilitates building a CSS string by appending rule blocks one at a time, checking whether the media query,
